feat(#45): move isAncestor, isDroppable to utils from hooks
